Consolidate shelters/permissions.py: move ShelterPermissions to accounts/types.py, teach make_granted_permissions to accept PermissionSet
