+++ apps/betterangels-backend/schema.graphql
@@ -2325,6 +2325,8 @@ enum ShelterPermissions {
   CHANGE
   DELETE
   VIEW
+  CHANGE_IS_REVIEWED
+  VIEW_PRIVATE
 }
 
 type ShelterPhotoType {
